Commit 3.12 - Deleting Node class
